hwbench/monitoring: Adding MonitoringDevice class
Future work to add different type of monitoring devices like PDUs, CDUs,
etc.. requires a generic class to share some behavior.

This commit is moving parts of the BMC class into the generic
MonitoringDevice one.

It also creates a new find_sections() function to make the lookup of
given type of MonitoringDevice inside the configuration file.
